Split out definition of statuses struct. Break circular header dependencies.
